+++ ldpd/log.h
@@ -16,50 +16,32 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
 
-#ifndef _LOG_H_
-#define	_LOG_H_
+#ifndef LOG_H
+#define LOG_H
 
 #include <stdarg.h>
 
-struct in6_addr;
-union ldpd_addr;
-struct hello_source;
-struct fec;
+extern const char	*log_procname;
 
-void		 logit(int, const char *, ...)
-			__attribute__((__format__ (printf, 2, 3)));
-void		 log_warn(const char *, ...)
-			__attribute__((__format__ (printf, 1, 2)));
-void		 log_warnx(const char *, ...)
-			__attribute__((__format__ (printf, 1, 2)));
-void		 log_info(const char *, ...)
-			__attribute__((__format__ (printf, 1, 2)));
-void		 log_notice(const char *, ...)
-			__attribute__((__format__ (printf, 1, 2)));
-void		 log_debug(const char *, ...)
-			__attribute__((__format__ (printf, 1, 2)));
-void		 fatal(const char *)
-			__attribute__ ((noreturn))
-			__attribute__((__format__ (printf, 1, 0)));
-void		 fatalx(const char *)
-			__attribute__ ((noreturn))
-			__attribute__((__format__ (printf, 1, 0)));
-const char	*log_sockaddr(void *);
-const char	*log_in6addr(const struct in6_addr *);
-const char	*log_in6addr_scope(const struct in6_addr *, unsigned int);
-const char	*log_addr(int, const union ldpd_addr *);
-char		*log_label(uint32_t);
-const char	*log_time(time_t);
-char		*log_hello_src(const struct hello_source *);
-const char	*log_map(const struct map *);
-const char	*log_fec(const struct fec *);
-const char	*af_name(int);
-const char	*socket_name(int);
-const char	*nbr_state_name(int);
-const char	*if_state_name(int);
-const char	*if_type_name(enum iface_type);
-const char	*msg_name(uint16_t);
-const char	*status_code_name(uint32_t);
-const char	*pw_type_name(uint16_t);
+void	 logit(int, const char *, ...)
+		__attribute__((__format__ (printf, 2, 3)));
+void	 vlog(int, const char *, va_list)
+		__attribute__((__format__ (printf, 2, 0)));
+void	 log_warn(const char *, ...)
+		__attribute__((__format__ (printf, 1, 2)));
+void	 log_warnx(const char *, ...)
+		__attribute__((__format__ (printf, 1, 2)));
+void	 log_info(const char *, ...)
+		__attribute__((__format__ (printf, 1, 2)));
+void	 log_notice(const char *, ...)
+		__attribute__((__format__ (printf, 1, 2)));
+void	 log_debug(const char *, ...)
+		__attribute__((__format__ (printf, 1, 2)));
+void	 fatal(const char *)
+		__attribute__ ((noreturn))
+		__attribute__((__format__ (printf, 1, 0)));
+void	 fatalx(const char *)
+		__attribute__ ((noreturn))
+		__attribute__((__format__ (printf, 1, 0)));
 
-#endif /* _LOG_H_ */
+#endif /* LOG_H */
